Add files via upload
It's a complete Restaurant Point of Sale System developed using WPF (Windows Presentation Foundation) framework of .NET Core with Xaml & C#, and ADO.NET for MSSQL Database Manipulation.

It let users select items from given panels either a general one or a category-specific. Then that item is added to the list of items and its price is displayed on a grid, along with the total and taxes on another grid. Users have the option either to remove a specific item, remove all items or generate a bill. If the user opts to generate a bill, then a pdf file is generated using the spite.PDF library and is saved in a specific folder, and all items are cleared for the new bill again.

NOTE: When you download and use it, don't forget to change the database connection string.